fix: int 나눗셈 -> double 나눗셈으로 로직 수정
- int 로 파라미터를 받아 둘을 나눠서 int 결과가 나와버림
- 문제에서 파라미터의 타입을 정수로 두라했으니 나누기할 때 하나만 double 로 바꿔서 나누기

+++ src/main/java/com/mosparta/calculator/CalculatorStep1.java
@@ -37,7 +37,7 @@ public class CalculatorStep1 {
                         break;
                     case '/':
                         if (num2 != 0) {
-                            System.out.println("Result: " + (num1 / num2));
+                            System.out.println("Result: " + ((double) num1 / num2));
                         } else {
                             System.out.println("Error: 0으로 나눌 수 없습니다");
                         }
